backend/requirements.txt
fastapi
uvicorn[standard]
PyJWT
passlib[bcrypt]
mysql-connector-python
redis
python-dotenv
python-multipart
kiteconnect
pandas
requests